+++ src/ResourceManager/Network/Commands.Network/VirtualNetworkGateway/RemoveAzureRmVpnClientIpsecParametersCommand.cs
@@ -32,6 +32,7 @@ namespace Microsoft.Azure.Commands.Network
             ParameterSetName = ParameterSetNames.ByFactoryName,
             Mandatory = true,
             HelpMessage = "The virtual network gateway name.")]
+        [ResourceNameCompleter("Microsoft.Network/virtualNetworkGateways", "ResourceGroupName")]
         [ValidateNotNullOrEmpty]
         public virtual string VirtualNetworkGatewayName { get; set; }