+++ src/ResourceManager/RecoveryServices/Commands.RecoveryServices.SiteRecovery.Test/ScenarioTests/A2A/AsrA2ATestsBase.cs
@@ -14,19 +14,16 @@
 
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
-using System.Net.Security;
-using System.Security.Cryptography.X509Certificates;
 using Microsoft.Azure.Commands.Common.Authentication;
 using Microsoft.Azure.Management.Internal.Resources;
 using Microsoft.Azure.Management.RecoveryServices;
 using Microsoft.Azure.Management.RecoveryServices.SiteRecovery;
-using Microsoft.Azure.Test;
 using Microsoft.Azure.Test.HttpRecorder;
 using Microsoft.WindowsAzure.Commands.ScenarioTest;
 using Microsoft.WindowsAzure.Commands.Test.Utilities.Common;
 using RestTestFramework = Microsoft.Rest.ClientRuntime.Azure.TestFramework;
-using Microsoft.Azure.Internal.Subscriptions;
 using Microsoft.Azure.ServiceManagemenet.Common.Models;
 using Microsoft.Azure.Commands.Common.Compute.Version_2018_04;
 
@@ -34,45 +31,31 @@ namespace RecoveryServices.SiteRecovery.Test
 {
     public abstract class AsrA2ATestsBase : RMTestBase
     {
-        protected string vaultSettingsFilePath;
-        protected string powershellFile;
-        protected string powershellHelperFile;
-        private CSMTestEnvironmentFactory csmTestFactory;
-        private EnvironmentSetupHelper helper;
+        protected string VaultSettingsFilePath;
+        protected string PowershellFile;
+        protected string PowershellHelperFile;
+        private EnvironmentSetupHelper _helper;
 
-        protected AsrA2ATestsBase()
+        protected void Initialize()
         {
-        }
-
-        protected void initialize()
-        {
-            this.helper = new EnvironmentSetupHelper();
+            _helper = new EnvironmentSetupHelper();
         }
 
         public ResourceManagementClient RmRestClient { get; private set; }
-        
-        
         public RecoveryServicesClient RecoveryServicesMgmtClient { get; private set; }
         public SiteRecoveryManagementClient SiteRecoveryMgmtClient { get; private set; }
-
         public ComputeManagementClient ComputeManagementClient { get; private set; }
-        public Microsoft.Azure.Management.Internal.Resources.ResourceManagementClient ResourceManagementRestClient { get; private set; }
+        public ResourceManagementClient ResourceManagementRestClient { get; private set; }
 
-        public void RunPowerShellTest(
-             XunitTracingInterceptor logger,
-             string scenario,
-            params string[] scripts)
+        public void RunPowerShellTest(XunitTracingInterceptor logger, string scenario, params string[] scripts)
         {
-            var callingClassType = TestUtilities.GetCallingClass(2);
-            var mockName = TestUtilities.GetCurrentMethodName(2);
+            var sf = new StackTrace().GetFrame(1);
+            var callingClassType = sf.GetMethod().ReflectedType?.ToString();
+            var mockName = sf.GetMethod().Name;
 
-            this.RunPsTestWorkflow(
+            RunPsTestWorkflow(
                 scenario,
                 () => scripts,
-
-                // no custom initializer
-                null,
-
                 // no custom cleanup 
                 null,
                 callingClassType,
@@ -82,72 +65,55 @@ namespace RecoveryServices.SiteRecovery.Test
         public void RunPsTestWorkflow(
             string scenario,
             Func<string[]> scriptBuilder,
-            Action<CSMTestEnvironmentFactory> initialize,
             Action cleanup,
             string callingClassType,
             string mockName)
         {
-            var providers = new Dictionary<string, string>();
-            providers.Add("Microsoft.Resources", null);
-            providers.Add("Microsoft.Features", null);
-            providers.Add("Microsoft.Authorization", null);
-
-            var providersToIgnore = new Dictionary<string, string>();
-            providersToIgnore.Add("Microsoft.Azure.Management.Internal.Resources.ResourceManagementClient", "2016-09-01");
-
-            HttpMockServer.Matcher = new PermissiveRecordMatcherWithApiExclusion(
-                true,
-                providers,
-                providersToIgnore);
-
-            HttpMockServer.RecordsDirectory = Path.Combine(
-                AppDomain.CurrentDomain.BaseDirectory,
-                "SessionRecords");
-
-            using (var context = RestTestFramework.MockContext.Start(
-                callingClassType,
-                mockName))
+            var providers = new Dictionary<string, string>
             {
-                this.csmTestFactory = new CSMTestEnvironmentFactory();
+                {"Microsoft.Resources", null},
+                {"Microsoft.Features", null},
+                {"Microsoft.Authorization", null}
+            };
 
-                if (initialize != null)
-                {
-                    initialize.Invoke(this.csmTestFactory);
-                }
+            var providersToIgnore = new Dictionary<string, string>
+            {
+                {"Microsoft.Azure.Management.Internal.Resources.ResourceManagementClient", "2016-09-01"}
+            };
 
-                this.SetupManagementClients(context);
+            HttpMockServer.Matcher = new PermissiveRecordMatcherWithApiExclusion(true, providers, providersToIgnore);
+            HttpMockServer.RecordsDirectory = Path.Combine( AppDomain.CurrentDomain.BaseDirectory, "SessionRecords");
 
-                this.helper.SetupEnvironment(AzureModule.AzureResourceManager);
+            using (var context = RestTestFramework.MockContext.Start(callingClassType, mockName))
+            {
+                SetupManagementClients(context);
 
-                var rmProfileModule = this.helper.RMProfileModule;
-                this.helper.SetupModules(
+                _helper.SetupEnvironment(AzureModule.AzureResourceManager);
+
+                var rmProfileModule = _helper.RMProfileModule;
+                _helper.SetupModules(
                     AzureModule.AzureResourceManager,
-                    powershellFile,
-                    powershellHelperFile,
+                    PowershellFile,
+                    PowershellHelperFile,
                     rmProfileModule,
-                    this.helper.RMResourceModule,
-                    this.helper.GetRMModulePath("AzureRM.RecoveryServices.psd1"),
-                    this.helper.GetRMModulePath("AzureRM.RecoveryServices.SiteRecovery.psd1"),
-                    "AzureRm.Resources.ps1");
+                    _helper.RMResourceModule,
+                    _helper.GetRMModulePath("AzureRM.RecoveryServices.psd1"),
+#if !NETSTANDARD
+                    _helper.GetRMModulePath("AzureRM.RecoveryServices.SiteRecovery.psd1"),
+#endif
+                    "AzureRM.Resources.ps1");
 
                 try
                 {
-                    if (scriptBuilder != null)
+                    var psScripts = scriptBuilder?.Invoke();
+                    if (psScripts != null)
                     {
-                        var psScripts = scriptBuilder();
-
-                        if (psScripts != null)
-                        {
-                            this.helper.RunPowerShellTest(psScripts);
-                        }
+                        _helper.RunPowerShellTest(psScripts);
                     }
                 }
                 finally
                 {
-                    if (cleanup != null)
-                    {
-                        cleanup.Invoke();
-                    }
+                    cleanup?.Invoke();
                 }
             }
         }
@@ -155,68 +121,36 @@ namespace RecoveryServices.SiteRecovery.Test
         protected void SetupManagementClients(
             RestTestFramework.MockContext context)
         {
-            this.RmRestClient = this.GetRmRestClient(context);
-            this.ResourceManagementRestClient = this.GetResourceManagementClientRestClient(context);
-            this.RecoveryServicesMgmtClient = this.GetRecoveryServicesManagementClient(context);
-            this.SiteRecoveryMgmtClient = this.GetSiteRecoveryManagementClient(context);
+            RmRestClient = GetRmRestClient(context);
+            ResourceManagementRestClient = GetResourceManagementClientRestClient(context);
+            RecoveryServicesMgmtClient = GetRecoveryServicesManagementClient(context);
+            SiteRecoveryMgmtClient = GetSiteRecoveryManagementClient(context);
             
-            this.helper.SetupManagementClients(
-                this.RmRestClient,
-                this.RecoveryServicesMgmtClient,
-                this.SiteRecoveryMgmtClient,
-                this.ResourceManagementRestClient);
+            _helper.SetupManagementClients(
+                RmRestClient,
+                RecoveryServicesMgmtClient,
+                SiteRecoveryMgmtClient,
+                ResourceManagementRestClient);
         }
 
-        private RecoveryServicesClient GetRecoveryServicesManagementClient(
-            RestTestFramework.MockContext context)
+        private static RecoveryServicesClient GetRecoveryServicesManagementClient(RestTestFramework.MockContext context)
         {
-            return context.GetServiceClient<RecoveryServicesClient>(
-                RestTestFramework.TestEnvironmentFactory.GetTestEnvironment());
+            return context.GetServiceClient<RecoveryServicesClient>(RestTestFramework.TestEnvironmentFactory.GetTestEnvironment());
         }
 
-        private ResourceManagementClient GetRmRestClient(
-            RestTestFramework.MockContext context)
+        private static ResourceManagementClient GetRmRestClient(RestTestFramework.MockContext context)
         {
-            return context.GetServiceClient<ResourceManagementClient>(
-                RestTestFramework.TestEnvironmentFactory.GetTestEnvironment());
+            return context.GetServiceClient<ResourceManagementClient>(RestTestFramework.TestEnvironmentFactory.GetTestEnvironment());
         }
 
-        private ResourceManagementClient GetResourceManagementClient(
-            RestTestFramework.MockContext context)
+        private static ResourceManagementClient GetResourceManagementClientRestClient(RestTestFramework.MockContext context)
         {
-            return context.GetServiceClient<ResourceManagementClient>(
-                RestTestFramework.TestEnvironmentFactory.GetTestEnvironment());
+            return context.GetServiceClient<ResourceManagementClient>(RestTestFramework.TestEnvironmentFactory.GetTestEnvironment());
         }
 
-        private Microsoft.Azure.Management.Internal.Resources.ResourceManagementClient GetResourceManagementClientRestClient(
-            RestTestFramework.MockContext context)
+        private static SiteRecoveryManagementClient GetSiteRecoveryManagementClient(RestTestFramework.MockContext context)
         {
-            return context.GetServiceClient<Microsoft.Azure.Management.Internal.Resources.ResourceManagementClient>(
-                RestTestFramework.TestEnvironmentFactory.GetTestEnvironment());
-        }
-
-        private SiteRecoveryManagementClient GetSiteRecoveryManagementClient(
-            string scenario,
-            RestTestFramework.MockContext context)
-        {
-            return context.GetServiceClient<SiteRecoveryManagementClient>(
-                RestTestFramework.TestEnvironmentFactory.GetTestEnvironment());
-        }
-
-        private SiteRecoveryManagementClient GetSiteRecoveryManagementClient(
-            RestTestFramework.MockContext context)
-        {
-            return context.GetServiceClient<SiteRecoveryManagementClient>(
-                RestTestFramework.TestEnvironmentFactory.GetTestEnvironment());
-        }
-        
-        private static bool IgnoreCertificateErrorHandler(
-            object sender,
-            X509Certificate certificate,
-            X509Chain chain,
-            SslPolicyErrors sslPolicyErrors)
-        {
-            return true;
+            return context.GetServiceClient<SiteRecoveryManagementClient>(RestTestFramework.TestEnvironmentFactory.GetTestEnvironment());
         }
     }
 }