+++ src/ResourceManager/Aks/Commands.Aks/ChangeLog.md
@@ -18,6 +18,13 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Minor changes for upcoming AzureRM to Az transition
+
+## Version 0.0.7
+* Update dependencies for type mapping issue
+
+## Version 0.0.6
+* Update version of YamlDotNet.Signed used
 
 ## Version 0.0.5
 * Updated to the latest version of the Azure ClientRuntime.