+++ src/Storage/Commands.Storage/ChangeLog.md
@@ -18,6 +18,14 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fix issue that Copy File cmdlet can't reuse source context in destination when not input -DestContext
+    - Start-AzureStorageFileCopy
+* Support Static Website configuration
+    - Enable-AzureStorageStaticWebsite
+    - Disable-AzureStorageStaticWebsite
+* Minor changes for upcoming AzureRM to Az transition
+
+## Version 4.6.1
 * Fix Copy Blob/File won't copy metadata when destination has metadata issue
     - Start-AzureStorageBlobCopy
     - Start-AzureStorageFileCopy